add constraints on user and node assigment to resource pools

+++ src/maasserver/models/resourcepool.py
@@ -21,6 +21,8 @@ from maasserver import DefaultMeta
 from maasserver.fields import MODEL_NAME_VALIDATOR
 from maasserver.models.cleansave import CleanSave
 from maasserver.models.timestampedmodel import TimestampedModel
+from maasserver.worker_user import user_name as worker_username
+from metadataserver.nodeinituser import user_name as node_init_username
 
 # The default pool cannot be renamed or deleted.
 DEFAULT_RESOURCEPOOL_NAME = 'default'
@@ -44,8 +46,16 @@ class ResourcePoolManager(Manager):
         return pool
 
     def get_user_resource_pools(self, user):
-        """Return ResourcePools a user has access to."""
+        """Return ResourcePools a User has access to."""
         return self.model.objects.filter(role__users=user)
+
+    def user_can_access_pool(self, user, pool):
+        """Whether a User has access to a ResourcePool."""
+        # maas internal users need to be able to own any node
+        admin_usernames = (worker_username, node_init_username)
+        if user.username in admin_usernames:
+            return True
+        return self.get_user_resource_pools(user).filter(pk=pool.pk).exists()
 
 
 class ResourcePool(CleanSave, TimestampedModel):
@@ -73,7 +83,41 @@ class ResourcePool(CleanSave, TimestampedModel):
         if self.is_default():
             raise ValidationError(
                 'This is the default pool, it cannot be deleted.')
+        if self.node_set.exists():
+            raise ValidationError(
+                'Pool has machines in it, it cannot be deleted.')
+        self._get_pool_role().delete()
         super().delete()
+
+    def grant_user(self, user):
+        """Grant user access to the resource pool.
+
+        XXX This should be dropped once we implement full RBAC, and the logic
+        moved to methods in Role.
+        """
+        role = self._get_pool_role()
+        role.users.add(user)
+
+    def revoke_user(self, user):
+        """Revoke user access to the resource pool.
+
+        XXX This should be dropped once we implement full RBAC, and the logic
+        moved to methods in Role.
+        """
+        from maasserver.models.node import Machine
+        if Machine.objects.filter(pool=self, owner=user).exists():
+            raise ValidationError(
+                'User has machines in the pool, it cannot be revoked.')
+        role = self._get_pool_role()
+        role.users.remove(user)
+
+    def _get_pool_role(self):
+        """Return the Role associated to the pool.
+
+        Until full RBAC is implemented, each ResourcePool is assigned to a
+        single role.
+        """
+        return self.role_set.first()
 
 
 # When a resource pool is created is created, create a default role associated
